change doubles to floats

+++ src/fishy.h
@@ -5,13 +5,13 @@
 
 #include <stdio.h>
 
-typedef double (*RHSFunc1D)(double x);
-typedef double (*RHSFunc2D)(double x, double y);
+typedef float (*RHSFunc1D)(float x);
+typedef float (*RHSFunc2D)(float x, float y);
 
 typedef struct {
-  double upper;
-  double diag;
-  double lower;
+  float upper;
+  float diag;
+  float lower;
 } TridiagMat;
 
 typedef struct {
@@ -20,22 +20,22 @@ typedef struct {
   TridiagMat lower;
 } BlockTridiagMat;
 
-void solve_tridiag_jacobi(TridiagMat A_h, double *rhs_values, int n, double *sol);
-void solve_tridiag_gs(TridiagMat A_h, double *rhs_values, int n, double *sol);
-void solve_blocktridiag_gs(BlockTridiagMat A_h, double *rhs_values, int n, double *sol);
+void solve_tridiag_jacobi(TridiagMat A_h, float *rhs_values, int n, float *sol);
+void solve_tridiag_gs(TridiagMat A_h, float *rhs_values, int n, float *sol);
+void solve_blocktridiag_gs(BlockTridiagMat A_h, float *rhs_values, int n, float *sol);
 
-void solve_poisson1d(double a, double b, double alpha, double beta, int n, RHSFunc1D f_rhs, double *sol) {
-  double h = (b - a) / (n + 1);
-  double *rhs_values = malloc((n + 2) * sizeof(double));
+void solve_poisson1d(float a, float b, float alpha, float beta, int n, RHSFunc1D f_rhs, float *sol) {
+  float h = (b - a) / (n + 1);
+  float *rhs_values = malloc((n + 2) * sizeof(float));
   rhs_values[0] = alpha;
   rhs_values[n+1] = beta;
 
   for (int i = 1; i < n + 1; i++) {
-    double x = a + i * h;
+    float x = a + i * h;
     rhs_values[i] = f_rhs(x);
   }
 
-  double inv_h_sq = 1 / (h*h);
+  float inv_h_sq = 1 / (h*h);
   TridiagMat A_h = {
     .upper = -1 * inv_h_sq,
     .diag = 2 * inv_h_sq,
@@ -47,21 +47,21 @@ void solve_poisson1d(double a, double b, double alpha, double beta, int n, RHSFu
   free(rhs_values);
 }
 
-void solve_poisson2d(double a, double b, double c, double d, int n, RHSFunc2D f_rhs, double *sol) {
-  double h1 = (b - a) / (n + 1);
-  double h2 = (d - c) / (n + 1);
-  double *rhs_values = malloc(n * n * sizeof(double));
+void solve_poisson2d(float a, float b, float c, float d, int n, RHSFunc2D f_rhs, float *sol) {
+  float h1 = (b - a) / (n + 1);
+  float h2 = (d - c) / (n + 1);
+  float *rhs_values = malloc(n * n * sizeof(float));
 
   for (int i = 0; i < n; i++) {
     for (int j = 0; j < n; j++) {
-      double x = a + (i + 1) * h1;
-      double y = c + (j + 1) * h2;
+      float x = a + (i + 1) * h1;
+      float y = c + (j + 1) * h2;
       rhs_values[IDX(i, j, n)] = f_rhs(x, y);
     }
   }
 
-  double inv_h1_sq = 1 / (h1*h1);
-  double inv_h2_sq = 1 / (h2*h2);
+  float inv_h1_sq = 1 / (h1*h1);
+  float inv_h2_sq = 1 / (h2*h2);
   TridiagMat B_tilde = {
     .upper = -1 * inv_h1_sq,
     .diag = 2 * inv_h1_sq + 2 * inv_h2_sq,
@@ -85,18 +85,18 @@ void solve_poisson2d(double a, double b, double c, double d, int n, RHSFunc2D f_
   free(rhs_values);
 }
 
-void solve_tridiag_jacobi(TridiagMat A_h, double *rhs_values, int n, double *sol) {
+void solve_tridiag_jacobi(TridiagMat A_h, float *rhs_values, int n, float *sol) {
   const int MAX_ITER = 20;
-  double *x_new = malloc(n * sizeof(double));
-  for (int i = 0; i < n; i++) sol[i] = 0.0;
+  float *x_new = malloc(n * sizeof(float));
+  for (int i = 0; i < n; i++) sol[i] = 0.0f;
 
-  double inv_diag = 1 / A_h.diag;
+  float inv_diag = 1 / A_h.diag;
 
   for (int iter = 0; iter < MAX_ITER; iter++) {
     // x_(k+1) = D^-1 (b - (L + U)x_k)
     for (int i = 0; i < n; i++) {
-      double left = (i == 0) ? 0.0 : sol[i-1];
-      double right = (i == n - 1) ? 0.0 : sol[i+1];
+      float left = (i == 0) ? 0.0f : sol[i-1];
+      float right = (i == n - 1) ? 0.0f : sol[i+1];
       x_new[i] = inv_diag * (rhs_values[i] - A_h.lower * left - A_h.upper * right);
     }
     for (int i = 0; i < n; i++) sol[i] = x_new[i];
@@ -104,39 +104,39 @@ void solve_tridiag_jacobi(TridiagMat A_h, double *rhs_values, int n, double *sol
   free(x_new);
 }
 
-void solve_tridiag_gs(TridiagMat A_h, double *rhs_values, int n, double *sol) {
+void solve_tridiag_gs(TridiagMat A_h, float *rhs_values, int n, float *sol) {
   const int MAX_ITER = 5000;
   for (int i = 0; i < n+2; i++) sol[i] = 0.0;
   sol[0] = rhs_values[0];
   sol[n+1] = rhs_values[n+1];
 
-  double inv_diag = 1 / A_h.diag;
+  float inv_diag = 1 / A_h.diag;
 
   for (int iter = 0; iter < MAX_ITER; iter++) {
     // x_(k+1) = (L + D)^-1 (b - U x_k)
     for (int i = 1; i < n+1; i++) {
-      double left = sol[i-1];
-      double right = sol[i+1];
+      float left = sol[i-1];
+      float right = sol[i+1];
       sol[i] = inv_diag * (rhs_values[i] - A_h.lower * left - A_h.upper * right);
     }
   }
 }
 
-void solve_blocktridiag_gs(BlockTridiagMat A_h, double *rhs_values, int n, double *sol) {
+void solve_blocktridiag_gs(BlockTridiagMat A_h, float *rhs_values, int n, float *sol) {
   const int MAX_ITER = 5000;
-  for (int k = 0; k < n * n; k++) sol[k] = 0.0;
+  for (int k = 0; k < n * n; k++) sol[k] = 0.0f;
 
-  double inv_diag = 1 / A_h.diag.diag;
+  float inv_diag = 1 / A_h.diag.diag;
 
   for (int iter = 0; iter < MAX_ITER; iter++) {
     // Gauss-Seidel over the 2D grid (lexicographic: j outer, i inner)
     for (int j = 0; j < n; j++) {
       for (int i = 0; i < n; i++) {
         int idx = IDX(i, j, n);
-        double left  = (i == 0)      ? 0.0 : sol[IDX(i - 1, j, n)];
-        double right = (i == n - 1)  ? 0.0 : sol[IDX(i + 1, j, n)];
-        double down  = (j == 0)      ? 0.0 : sol[IDX(i, j - 1, n)];
-        double up    = (j == n - 1)  ? 0.0 : sol[IDX(i, j + 1, n)];
+        float left  = (i == 0)      ? 0.0f : sol[IDX(i - 1, j, n)];
+        float right = (i == n - 1)  ? 0.0f : sol[IDX(i + 1, j, n)];
+        float down  = (j == 0)      ? 0.0f : sol[IDX(i, j - 1, n)];
+        float up    = (j == n - 1)  ? 0.0f : sol[IDX(i, j + 1, n)];
 
         sol[idx] = inv_diag * (
           rhs_values[idx]
